move constants from src to utils
